server/requirements.txt
Flask
paho-mqtt
pymongo
PyYAML
requests
